Use x86-specific source files only for x86 target
Previously, we used x86-specific source files (those for sse2)
regardless of the target because x86 was the only supported host target.
This becomes a problem as we add the support for the new ARM64 host
target - linux_bionic_arm64. Now, those source files are built only when
the target architecture is x86.

Bug: 159685774
Test: HOST_CROSS_OS=linux_bionic HOST_CROSS_ARCH=arm64 m \
out/soong/host/linux_bionic-arm64/bin/webRTC

Change-Id: Id13d66ccd5ae6f8596ce32d4ccb2735856d2159a